+++ src/main/kotlin/cloud/glitchdev/rfu/feature/mob/BossHealthBar.kt
@@ -3,22 +3,29 @@ package cloud.glitchdev.rfu.feature.mob
 import cloud.glitchdev.rfu.config.categories.SeaCreatureConfig
 import cloud.glitchdev.rfu.config.categories.SeaCreatureConfig.HEALTH_BAR_REGEX
 import cloud.glitchdev.rfu.events.managers.MobEvents.registerMobDetectEvent
+import cloud.glitchdev.rfu.events.managers.TickEvents.registerTickEvent
 import cloud.glitchdev.rfu.feature.Feature
 import cloud.glitchdev.rfu.feature.RFUFeature
 import cloud.glitchdev.rfu.gui.hud.elements.bossbar.BossHealthBarDisplay
-import cloud.glitchdev.rfu.data.mob.MobManager
 
 @RFUFeature
 object BossHealthBar : Feature {
     override fun onInitialize() {
         registerMobDetectEvent { entities ->
-            if(!SeaCreatureConfig.bossHealthBars) {
+            if (!SeaCreatureConfig.bossHealthBars) {
                 BossHealthBarDisplay.updateEntities(emptySet())
                 return@registerMobDetectEvent
             }
             val filteredEntities = entities.filter { HEALTH_BAR_REGEX.matches(it.sbName) }
             BossHealthBarDisplay.updateEntities(filteredEntities.toSet())
-            if(SeaCreatureConfig.boostPollingRate) MobManager.boostDetectionRate(filteredEntities.isNotEmpty())
+        }
+
+        registerTickEvent(interval = 1L) {
+            if (!SeaCreatureConfig.bossHealthBars || BossHealthBarDisplay.entities.isEmpty()) return@registerTickEvent
+
+            BossHealthBarDisplay.entities.removeIf { it.isRemoved() }
+            BossHealthBarDisplay.entities.forEach { it.updateEntityData() }
+            BossHealthBarDisplay.updateState()
         }
     }
 }