+++ CHANGELOG.md
@@ -10,5 +10,7 @@
 
 ### Changes
 - Changed the !since command to display only simple info
+- Removed Boost Polling rate setting and changed the display itself to update every 1 tick instead of all mobs
+updating every 2 ticks when with a rare sc
 
 ### Back-end